python/collect_data/requirements.txt
beautifulsoup4==4.10.0
docx2txt==0.8
pymupdf==1.19.1
fitz==0.0.1.dev2
GitPython==3.1.24
Markdown==3.3.4
pandas==1.2.4
python_pptx==0.6.21
Unidecode==1.3.2
